# PROYECTO-6 MONGO.postman_collection.json-OpenApi3Yaml.yaml
openapi: 3.0.0
info:
  title: PROYECTO-6 MONGO
  contact: {}
  version: '1.0'
servers:
- url: https://mongodb-proyecto6.onrender.com/api
  variables: {}
paths:
...
  /clients/update/679aac1e2a87b842e15b44ad:
    put:
      tags:
      - CLIENT
      summary: UPDATE CLIENT BY ID
      operationId: UPDATECLIENTBYID
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UPDATECLIENTBYIDREQUEST'
              - example:
                  password: marcelo
            example:
              password: marcelo
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /clients/client-login:
    post:
      tags:
      - CLIENT
      summary: CLIENT LOGIN
      operationId: CLIENTLOGIN
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CLIENTLOGINREQUEST'
              - example:
                  username: la_guada
                  password: '12345'
            example:
              username: la_guada
              password: '12345'
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    UPDATECLIENTBYIDREQUEST:
      title: UPDATECLIENTBYIDREQUEST
      required:
      - password
      type: object
      properties:
        password:
          type: string
      example:
        password: marcelo
    CLIENTLOGINREQUEST:
      title: CLIENTLOGINREQUEST
      required:
      - username
      - password
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      example:
        username: la_guada
        password: '12345'
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...